+++ code/packages/perl/radix-tree/t/00-load.t
@@ -1,0 +1,8 @@
+use strict;
+use warnings;
+use Test::More;
+
+ok(eval { require CodingAdventures::RadixTree; 1 }, 'CodingAdventures::RadixTree loads');
+ok(CodingAdventures::RadixTree->VERSION, 'has a VERSION');
+
+done_testing;